src/test/javascript/spec/app/entities/historique-appel/historique-appel.service.spec.ts
/* tslint:disable max-line-length */
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { HttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { take, map } from 'rxjs/operators';
import * as moment from 'moment';
import { DATE_FORMAT } from 'app/shared/constants/input.constants';
import { HistoriqueAppelService } from 'app/entities/historique-appel/historique-appel.service';
import { IHistoriqueAppel, HistoriqueAppel } from 'app/shared/model/historique-appel.model';

describe('Service Tests', () => {
    describe('HistoriqueAppel Service', () => {
        let injector: TestBed;
        let service: HistoriqueAppelService;
        let httpMock: HttpTestingController;
        let elemDefault: IHistoriqueAppel;
        let currentDate: moment.Moment;
        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [HttpClientTestingModule]
            });
            injector = getTestBed();
            service = injector.get(HistoriqueAppelService);
            httpMock = injector.get(HttpTestingController);
            currentDate = moment();

            elemDefault = new HistoriqueAppel(0, currentDate, 'AAAAAAA', 'AAAAAAA');
        });

        describe('Service methods', async () => {
            it('should find an element', async () => {
                const returnedFromService = Object.assign(
                    {
                        dateAppel: currentDate.format(DATE_FORMAT)
                    },
                    elemDefault
                );
                service
                    .find(123)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: elemDefault }));

                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should create a HistoriqueAppel', async () => {
                const returnedFromService = Object.assign(
                    {
                        id: 0,
                        dateAppel: currentDate.format(DATE_FORMAT)
                    },
                    elemDefault
                );
                const expected = Object.assign(
                    {
                        dateAppel: currentDate
                    },
                    returnedFromService
                );
                service
                    .create(new HistoriqueAppel(null))
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'POST' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should update a HistoriqueAppel', async () => {
                const returnedFromService = Object.assign(
                    {
                        dateAppel: currentDate.format(DATE_FORMAT),
                        description: 'BBBBBB',
                        nomClient: 'BBBBBB'
                    },
                    elemDefault
                );

                const expected = Object.assign(
                    {
                        dateAppel: currentDate
                    },
                    returnedFromService
                );
                service
                    .update(expected)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'PUT' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should return a list of HistoriqueAppel', async () => {
                const returnedFromService = Object.assign(
                    {
                        dateAppel: currentDate.format(DATE_FORMAT),
                        description: 'BBBBBB',
                        nomClient: 'BBBBBB'
                    },
                    elemDefault
                );
                const expected = Object.assign(
                    {
                        dateAppel: currentDate
                    },
                    returnedFromService
                );
                service
                    .query(expected)
                    .pipe(
                        take(1),
                        map(resp => resp.body)
                    )
                    .subscribe(body => expect(body).toContainEqual(expected));
                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify([returnedFromService]));
                httpMock.verify();
            });

            it('should delete a HistoriqueAppel', async () => {
                const rxPromise = service.delete(123).subscribe(resp => expect(resp.ok));

                const req = httpMock.expectOne({ method: 'DELETE' });
                req.flush({ status: 200 });
            });
        });

        afterEach(() => {
            httpMock.verify();
        });
    });
});
